Crate sgx_build_helper

Macros

t
A helper macro to unwrap a result except also print out details like:

Structs

NativeLibBoilerplate

Functions

gnu_target

make

mtime
Returns the last-modified time for path, or zero if it doesn't exist.

native_lib_boilerplate

output

rerun_if_changed_anything_in_dir

run

run_silent

run_suppressed

sanitizer_lib_boilerplate

try_run_silent

try_run_suppressed

up_to_date
Returns whether dst is up to date given that the file or files in src are used to generate it.